refactor: refactor css vars and interface
