+++ packages/zarm/src/index.ts
@@ -114,7 +114,12 @@ export { default as Scroller } from './scroller';
 export type { SearchBarProps } from './search-bar';
 export { default as SearchBar } from './search-bar';
 
-export type { SkeletonProps, SkeletonTitleProps, SkeletonParagraphProps } from './skeleton';
+export type {
+  SkeletonProps,
+  SkeletonTitleProps,
+  SkeletonParagraphProps,
+  SkeletonCssVars,
+} from './skeleton';
 export { default as Skeleton } from './skeleton';
 
 export type { SliderProps, SliderCssVars } from './slider';
