+++ packages/zarm/src/skeleton/demo.md
@@ -95,18 +95,18 @@ ReactDOM.render(<Demo />, mountNode);
 
 ### Skeleton
 
-| 属性            | 默认值                      | 说明 |
-| :-------------- | :-------------------------- | :--- |
-| --width         | -                           | 宽度 |
-| --height        | 'var(--za-skeleton-height)' | 高度 |
-| --border-radius | 'var(--za-skeleton-radius)' | 圆角 |
-| --margin        | 'var(--za-skeleton-margin)' | 边距 |
+| 属性            | 默认值   | 说明 |
+| :-------------- | :----- | :--- |
+| --width         | -      | 宽度 |
+| --height        | '18px' | 高度 |
+| --border-radius | '2px'  | 圆角 |
+| --margin        | '0'    | 边距 |
 
 ### Skeleton.Title
 
-| 属性            | 默认值                            | 说明 |
-| :-------------- | :-------------------------------- | :--- |
-| --width         | 'var(--za-skeleton-title-width)'  | 宽度 |
-| --height        | 'var(--za-skeleton-title-height)' | 高度 |
-| --border-radius | 'var(--za-skeleton-radius)'       | 圆角 |
-| --margin        | 'var(--za-skeleton-title-margin)' | 边距 |
+| 属性            | 默认值      | 说明 |
+| :-------------- | :-------- | :-- |
+| --width         | '45%'     | 宽度 |
+| --height        | '32px'    | 高度 |
+| --border-radius | '2px'     | 圆角 |
+| --margin        | '16px 0'  | 边距 |
